feature: create carousel component
