+++ client/src/view/components/Carousel/config/default.jsx
@@ -1,15 +1,18 @@
-const defaultSlide = {
+const slide = {
   className: 'swiper__slide'
 };
 
-const defauldCarousel = {
+const carousel = {
   className: 'header__carousel',
   loop: true,
   effect: 'cube',
   pagination: {
     clickable: true
   },
+  autoplay: {
+    delay: 5000
+  },
   navigation: true
 };
 
-export default { defaultSlide, defauldCarousel };
+export default { slide, carousel };
