+++ client/src/view/components/Carousel/index.jsx
@@ -2,5 +2,5 @@ import Carousel from './Carousel';
 
 export default Carousel;
 
-export { default as data } from './config/default';
+export { default as defaultCarousel } from './config/default';
 export { default as popularProductCard } from './config/custom';
