+++ client/src/view/pages/Home/Home.jsx
@@ -1,42 +1,14 @@
-import React, { useEffect, useState, useMemo } from 'react';
+import React from 'react';
+import HeaderCarousel from './components/HeaderCarousel';
 
-// import PropTypes from 'prop-types';
-import Carousel, { popularProductCard } from '../../components/Carousel';
-import { sliderRequests } from '../../../api/server';
-
-const Image = ({ src, className }) => {
-  const alt = src.split('/').pop().split('.')[0];
-  return <img src={src} className={className} alt={alt} />;
-};
+import './style.scss';
 
 const Home = () => {
-  const [images, setImages] = useState([]);
-
-  useEffect(async () => {
-    setImages(await sliderRequests.getSlides());
-  }, []);
-
-  console.log(images);
-  const imagesComp = useMemo(
-    () =>
-      images.map(({ imageUrl }, key) => (
-        <Image src={imageUrl} key={key} className="swiper__image" />
-      )),
-    [images]
-  );
-
-  const { slide, carousel } = popularProductCard;
-
   return (
-    <>
-      <Carousel
-        components={imagesComp}
-        slideProps={slide}
-        carouselProps={carousel}
-      />
-    </>
+    <div className="carousel-wrapper">
+      <HeaderCarousel />
+    </div>
   );
 };
-Home.propTypes = {};
 
 export default Home;
